backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
praw==7.7.1
textblob==0.17.1
curl-cffi==0.6.2
